add: group based on confidence
The "grouped" function in the evaluate.py in the SDK uses the mode to calculate the matching Annotation Set.
This does not work if the Tokenizer creates a large number of Spans that match correct Spans in the ground truth. Only because "correct" Spans are found, they should not be used for mode calculation as they have a confidence below label threshold.